move log content to log widget
